Update patch set 7
Patch Set 7:

Actually, it does on some points. Maybe I'm not using the right terms, as I agree it's not diverging in the sense "not compatible".
E.g: expiration is set to one day, while users can set it up to 7 days or 15 minutes by default.

"A request must reach AWS within 15 minutes of the
time stamp in the request. Otherwise, AWS denies the request." 
Page 17
https://d0.awsstatic.com/whitepapers/Security/Intro_Security_Practices.pdf

Patch-set: 7