Fix query string generator
The standard urlencode function does not work because of some ambiguity in the underlying RFC.
This patch use quote directly to control what char need to be encoded, and which one need to be escaped.

In python:

urlencode({"q":" "})
'q=+'

In the AWS doc:

The space character is a reserved character and must be encoded as "%20" (and not as "+").

Change-Id: I745286cd808ad1b0c75badc1f5fb2ef1267ff670